requirements.txt
python-dotenv
PyGithub
gitpython